# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mobile Messenger API",
    "version": "0.1.0",
    "description": "Canonical REST contract shared by the NestJS backend, web client, and iOS client."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/matrix/config": {
      "get": {
        "operationId": "getMatrixConfiguration",
        "tags": [
          "matrix"
        ],
        "description": "Disabled unless MATRIX_ENABLED=true. Metadata only; no messages, access tokens or cryptographic keys cross this boundary. Revoking any Nest user session conservatively revokes all Matrix sessions for that account.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatrixConfiguration"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Matrix disabled or chat unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Immutable binding or membership epoch conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Matrix revocation pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/matrix/chats/{chatID}": {
      "get": {
        "operationId": "getMatrixChatDescriptor",
        "tags": [
          "matrix"
        ],
        "description": "Disabled unless MATRIX_ENABLED=true. Metadata only; no messages, access tokens or cryptographic keys cross this boundary. Revoking any Nest user session conservatively revokes all Matrix sessions for that account.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatrixChatDescriptor"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Matrix disabled or chat unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Immutable binding or membership epoch conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Matrix revocation pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "chatID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "post": {
        "operationId": "bindMatrixChatRoom",
        "tags": [
          "matrix"
        ],
        "description": "Disabled unless MATRIX_ENABLED=true. Metadata only; no messages, access tokens or cryptographic keys cross this boundary. Revoking any Nest user session conservatively revokes all Matrix sessions for that account.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatrixChatDescriptor"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Matrix disabled or chat unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Immutable binding or membership epoch conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Matrix revocation pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "chatID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindMatrixRoom"
              }
            }
          }
        }
      }
    },
    "/api/matrix/revocation": {
      "get": {
        "operationId": "getMatrixRevocationStatus",
        "tags": [
          "matrix"
        ],
        "description": "Disabled unless MATRIX_ENABLED=true. Metadata only; no messages, access tokens or cryptographic keys cross this boundary. Revoking any Nest user session conservatively revokes all Matrix sessions for that account.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatrixRevocationStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Matrix disabled or chat unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Immutable binding or membership epoch conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Matrix revocation pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oidcBridgeSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-OIDC-Bridge-Secret",
        "description": "Internal server-to-server credential: canonical base64url encoding of 32 random bytes. Never place in iOS, web bundles, cookies, or user JWTs. Requires OIDC_BRIDGE_ENABLED=true and private network access."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "details": {}
        }
      },
...
      "MatrixConfiguration": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "homeserverURL",
          "issuerURL",
          "serverName",
          "webClientID",
          "expectedUserID"
        ],
        "properties": {
          "homeserverURL": {
            "type": "string",
            "format": "uri"
          },
          "issuerURL": {
            "type": "string",
            "format": "uri"
          },
          "serverName": {
            "type": "string"
          },
          "webClientID": {
            "type": "string"
          },
          "expectedUserID": {
            "type": "string"
          }
        }
      },
      "MatrixParticipant": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "userID",
          "matrixUserID",
          "displayName",
          "active"
        ],
        "properties": {
          "userID": {
            "type": "string",
            "format": "uuid"
          },
          "matrixUserID": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "MatrixChatDescriptor": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "chatID",
          "roomID",
          "epoch",
          "boundByUserID",
          "participants",
          "boundEpoch"
        ],
        "properties": {
          "chatID": {
            "type": "string",
            "format": "uuid"
          },
          "roomID": {
            "type": [
              "string",
              "null"
            ]
          },
          "epoch": {
            "type": "integer",
            "minimum": 1
          },
          "boundByUserID": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "participants": {
            "type": "array",
            "maxItems": 128,
            "items": {
              "$ref": "#/components/schemas/MatrixParticipant"
            }
          },
          "boundEpoch": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "description": "Membership epoch at immutable room registration. Clients must reject sends if it differs from epoch until a reviewed membership migration completes."
          }
        }
      },
      "BindMatrixRoom": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "roomID",
          "epoch"
        ],
        "properties": {
          "roomID": {
            "type": "string",
            "maxLength": 255,
            "description": "Opaque Matrix room ID, including room-version-12 IDs. Clients must verify encryption, membership and device trust with the official SDK; a binding does not attest room security."
          },
          "epoch": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "MatrixRevocationStatus": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "pending"
        ],
        "properties": {
          "pending": {
            "type": "boolean"
          }
        }
      }
    }
  }
}